Custom serializers
Work to get a lot of custom serializers in place

Kotlin singleton serializer

Almost working generics

Rebase complete and tidied a little

Much better support for generics and custom serializers.  Still need to work on schema for generics.

Still work in progress, but opted out the AMQP serialization in Kryo.